一些小的改动

+++ src/com/yuanchuang/yohey/view/MyImageView.java
@@ -1,8 +1,6 @@
 package com.yuanchuang.yohey.view;
 
 import com.yuanchuang.yohey.R;
-import com.yuanchuang.yohey.R.styleable;
-
 import android.content.Context;
 import android.content.res.TypedArray;
 import android.graphics.Bitmap;
@@ -229,10 +227,10 @@ public class MyImageView extends ImageView {
 		mShaderMatrix.set(null);
 
 		if (mBitmapWidth * mDrawableRect.height() > mDrawableRect.width() * mBitmapHeight) {
-			scale = mDrawableRect.height() / (float) mBitmapHeight;
+			scale = mDrawableRect.height() / mBitmapHeight;
 			dx = (mDrawableRect.width() - mBitmapWidth * scale) * 0.5f;
 		} else {
-			scale = mDrawableRect.width() / (float) mBitmapWidth;
+			scale = mDrawableRect.width() / mBitmapWidth;
 			dy = (mDrawableRect.height() - mBitmapHeight * scale) * 0.5f;
 		}
